backend/staticfiles/staticfiles/requirements.txt
asgiref==3.8.1
Django==4.2.7
django-cors-headers==3.14.0
django-jazzmin==3.0.1
django-shortuuidfield==0.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
inflection==0.5.1
packaging==24.1
pillow==11.0.0
PyJWT==2.6.0
pytz==2024.2
PyYAML==6.0.2
shortuuid==1.0.13
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
